ml/requirements.txt
fastapi
uvicorn
torch
transformers
celery
redis==5.0.1
duckduckgo-search==5.3.0
beautifulsoup4==4.12.3
chromadb==0.4.24
python-multipart
requests
sqlalchemy
asyncpg
pydantic-settings
alembic
python-dotenv
sentence-transformers
scikit-learn
pandas
captum
slowapi
